lab-5.2-autogen-incident-team/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pyautogen==0.2.27
redis==5.0.3
sqlalchemy==2.0.29
psycopg2-binary==2.9.9
pydantic==2.6.4
python-dotenv==1.0.1
requests==2.31.0
docker==7.0.0
prometheus-client==0.20.0
